Custom attributes : Allow to directly define priority, severity, reproducibility, status (with an associated color), resolution, projection and eta in Mantis instead of editing the config file.
Administrators make settings in Manage > Manage project (manage_proj_menu_page.php), and
this option can be desactivated in the config_inc1.php

For the time beings, custom attributes are effective for all the projects but implementation has been made with a 'per Project Settings' in mind, the main difficulty will lie in displaying the status in the view_all page. When this improvment (planned for > 0.18) is made, move the code to manage_proj_edit_page.php, it would be dealed almost like the categories.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@1182 f5dc347c-c33d-0410-90a0-b07cc1902cb9