Fixed a problem with victor's cleanup.
In versions of php less than 4.1.0, even though we've defined $_* they aren't auto-globals like there are after 4.1.0 so we still need to pull them in as globals.

I've been putting this inside conditionals because I don't think there are actually that many places that need to access them directly and this way we can easily grep for code that can be removed if we ever increase the minimum requirement.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@1314 f5dc347c-c33d-0410-90a0-b07cc1902cb9