And after much deliberation, lo, the minimum requirements were bumped to PHP 4.0.4 (released nearly two years ago) and there was much rejoicing (well maybe a little rejoicing anyway).
We may consider bumping it higher to 4.0.6 so we can use argument swapping in sprintf() to make internationalization easier... but we'll bump it if we get there.

Note, as a base line, RedHat 7.1 and up shipped with 4.0.6 or higher.  And all redhat 7.x versions have had rpm updates released to bring them up to 4.1.2 to fix security holes.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@1502 f5dc347c-c33d-0410-90a0-b07cc1902cb9