clean up bug_view_*.php a bit
They use the new api but they need a more thorough read-through... the CRC checking stuff is making the advanced page ugly at the moment... it's spread throughout the page... perhaps we could just have a different page for it?  It will be better when the new access code is rewritten since the access code can return true if the user is authorized or if there is a CRC present for the page.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@1599 f5dc347c-c33d-0410-90a0-b07cc1902cb9