Replace all instances of empty() with our new function called is_blank().
empty() is a horrid function which isn't intended for checking strings.  It is intended for checking whether a variable is unset or false (or rather anything that == false, which includes 0 and '').  In every case, as far as we can tell, we are not checking for this but rather for whether the string contains information.  So this function first performs trim() on the string and then checks whether the length is 0.

The patch was submitted by Robert Foster <rfoster@mountainvisions.com.au> at my request.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@1626 f5dc347c-c33d-0410-90a0-b07cc1902cb9