* core/API.php: add line to load new utility_api file * core/helper_api.php: remove some functions and move them to utility_api.php * core/project_api.php: rename uses of helper_terminate_directory_path() to terminate_directory_path() since it is no longer in the helper_api * core/utility_api.php: new file to hold utility functions needed by all APIs
I added a Utility API which contains little functions that need to be available to all APIs (and therefore do not have dependcies on any other APIs themselves).  These are the types of little functions that do little conversions or things that should be in php but aren't, etc.

This was needed to make the is_blank() change work, otherwise there was a an ugly dependency that pushed helper_api and print_api above nearly everything else.  I think helper_api needs to come after base APIs like DB and config, but before most other ones (as it is now).


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@1627 f5dc347c-c33d-0410-90a0-b07cc1902cb9