* doc/README.LDAP: correct a few typos using $t_ instead of $g_ * doc/configuration.html: change g_ldap_organisation to g_ldap_organization * config_defaults_inc.php   (g_ldap_bind_dn): add config variable   (g_ldap_bind_passwd): add config variable   (g_ldap_organisation): removed in favour of g_ldap_organization   (g_ldap_organization): replacing g_ldap_organisation for consistency   (g_ldap_port): add config variable * core/ldap_api.php: change g_ldap_organisation to g_ldap_organization * core/obsolete.php: add an entry to notify that g_ldap_organisation is obsolete and g_ldap_organization is now used
I hope this is right.  I originally was fixing the problem that Robert's new config options weren't initialized in the default config file.  Then I noticed the two typos in the README.  Then I noticed that our config option spelled organisation with an 's' but the LDAP field name (and most of the example of organizationName I could find relating to LDAP on Google spelled it with a 'z'.

If anyone objects, I can revert this.  I wouldn't normally rename a variable name because of an alternate spelling but this seems to be a case where we're not using what everyone else uses in this context.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@1641 f5dc347c-c33d-0410-90a0-b07cc1902cb9