Implemented #2773 when e-mail is sent by phpmailer.
Description
"Add an option to allow the admin to debug the e-mail feature by specifying a test e-mail account that should receive all e-mail messages. When this e-mail is supplied, the e-mail should only be sent to this account, while the original To, CC, and BCC should be included as part of the message body."

What is implemented
"This feature is now supported when phpmailer is used. Following are the details:
- If e-mail notifications is OFF, no e-mails will be sent at all, even if the debugging feature is ON.
- $g_debug_email can be set to OFF or an e-mail address. The default value is OFF. If it is not OFF, then the debug feature is activated, where all messages are only sent to the specified address, while including the original to, cc, bcc in the message body. "


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@1655 f5dc347c-c33d-0410-90a0-b07cc1902cb9