Finally, the new database upgrade stuff.  I would recommend those of you who aren't developers refrain from using it until a few of the developers have tested it to see if I've missed anything huge.  That way if we have to adjust the upgrade ids for any reason you won't be in a mixed state.
I would recommend testing this on a copy of your database.

But, I have tested installing 0.14, 0.15, 0.16, and 0.17 databases and all of them now update cleanly all the way and actually end up looking the same (which they didn't before).  These were all empty databases though, so I'm looking forward to some people running tests on copies of their actual databases with data.

The one problem I know of that still exists is that the dates seem to get updated during this process.  I haven't looked into why that happens yet, but there were workarounds for it in the old upgrade files.  We'll either need to solve it or add our own workarounds in.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@1740 f5dc347c-c33d-0410-90a0-b07cc1902cb9