load php_api fvery first thing in core.php again
Jeroen, this is an example of the bootstrapping problem I was talking about.  We can't include this before the config files without hardcoding the core/ path.  But if we don't include it before the config files, then old versions of PHP will break (for example on PHP4.0.6, the code to get the default hostname doesn't work because $_SERVER doesn't exist yet.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@1777 f5dc347c-c33d-0410-90a0-b07cc1902cb9