Fix #3064:  view bug buttons in vertical column
Many browsers go to the next line when you start a form.  The only way to ensure they're on the same line is to put them each in their own table cell.  If we want them to be close together and centered then they need to go into a table inside the main table cell.  There may be another way to do this with CSS but I'm not going to spend time on it right now.

* bug_view_advanced_page.php:
* bug_view_page.php:
  use align="center" instead of class="center"
  since the css version doesn't seem to work when the td has a table as its contents?

* core/html_api.php:
  (html_buttons_view_bug_page): put each button in its own td and put
  the whole thing inside a table.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@2125 f5dc347c-c33d-0410-90a0-b07cc1902cb9