Re-designed filter display/form.
The filter form on the bug list page looks a little more 'together' now (I hope). But even more importantly, the fields in the filter edit page are now in the same order as the filter display above the bug list page. So the user doesn't have to search the field anymore after he clicks on one of the filter edit links.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@2585 f5dc347c-c33d-0410-90a0-b07cc1902cb9