0007125: Application Error 200 0006633: Editing a bug in advanced bug update page
I can see this switch statement containing a mixture of path + params in the future.

For example, for a user profile page that for example showed 'issues reported by user in this project' we'd probably want to keep the param when changing project.

Hence whilst i've gone for a 'deny list', we could equally do the reverse and only allow certain urls as at the end, it's likely the length of the list would be the same.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@4510 f5dc347c-c33d-0410-90a0-b07cc1902cb9