The minimal php version for mantis 1.1 is PHP 4.3.0 (as per mailing list discussion).
nexen.net provides a survey of ~11 million webservers showing php versions. <5% of webservers run a version of php < 4.3 according to that site.

This commit removes the legacy support for $_GET/POST support for php versions where these are not globally defined.
On  my data set, this reduces 2700 calls to php_version_at_least to 1 call.


git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@4570 f5dc347c-c33d-0410-90a0-b07cc1902cb9