ADODB PGSQL DATADICT FIX:
a) AlterColumnSQL appears to pass the type instead of column name to the NULL/NOT NULL query
b) If you are converting from a BOOL->INT in sql, whilst select (1::INT4)::BOOL is valid, you can not cast in the ALTER TABLE syntax without a USING clause. In addition, it appears one needs to drop then readd any default value.

git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@4715 f5dc347c-c33d-0410-90a0-b07cc1902cb9