Fix up messages files header and make them consistent. Also add some doxygen tags in there.
This will make for easier parsing if a translation system is used. Left legacy files untouched, as they will probably be removed some time in the future anyway.

Question: should we add a DEPRECATED warning to the messages files with encodings other than UTF8, so someone will not start working on any of them by accident?

git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@5547 f5dc347c-c33d-0410-90a0-b07cc1902cb9