1. Only store columns configuration in database if it's different from default. (As config_api stores complex objects as serialized strings, we compare the new/old serialized string) 2. Modify config_defaults_inc.php to use category_id not category for category column as this seems to be what works.
Note:
A) should it be category or category_id? - whilst category_id seems to be correct, sorting fails.
B) columns API doesn't seem to be as generic as it could be -> we allow word export and also plugins to implement things like mindmap export, but dont provide a way for users to configure fields for these.

git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@5577 f5dc347c-c33d-0410-90a0-b07cc1902cb9