Create new g_bug_closed_status_threshold setting to allow for custom installations where there may exist different or multiple closed statuses (COMPLETED vs NOT COMPLETED, IMPLEMENTED, etc).
The summary page used to use hardcoded defaults such as CLOSED
and RESOLVED. This patch replaces those hardcoded defaults with
the thresholds set in the Mantis configuration. The summary page
will now show the number of resolved and closed tickets
correctly in a custom installation where the default RESOLVED
and or CLOSED statuses are not being used.

Signed-off-by: John Reese <jreese@leetcode.net>