Add Stick/Unstick button to html_buttons_view_bug_page
Currently there is no way to stick/unstick a bug when you're
viewing a bug. At present you need to go to View Issues,
select a checkbox for each bug you want to stick/unstick and
confirm those changes.

This patch adds a new Stick/Unstick button to the view bug
pages (both simple and advanced modes) next to the existing
Monitor Issue button. If the bug is currently stuck, the
button will read "Unstick". If the ticket is currently not
stuck, the button will read "Stick". Clicking on the button
will automatically perform the required stick/unstick
action without further confirmation, and will redirect the
user back to the bug view page.