Fix #10544: Fix bugnote order and limit when generating emails.
Reworked the original implementation of bugnote_get_all_bugnotes() to
not care about order or limits, and just cache all the bugnotes in a
simplified format.  Updated bugnote_get_visible_bugnotes() to now
expect a full list of bugnotes, and correctly handle the display order
and limit each time called.