Custom priority & severity significance thresholds
Previously the thresholds at which a priority or severity were
considered significant were hardcoded. Significant bugs are displayed or
printed with emphasis.

This patch introduces configuration options to allow the user to
manually set their own thresholds of significance. This is particularly
useful for custom environments where custom priorities and severities
are being used.

This patch also recognises that multiple closed statuses may exist.
Previously it was assumed that there was only ever one closed status.