Notify user if an admin updates their account
This fixes bug #4772 - a request for this new feature.

When managing user accounts, administrators now have a checkbox (checked
by default) which allows them to send notifications of account changes
to the current/new email address of the user.

This is particularly useful for informing users that the username of
their account has been updated. Previously, an email or other method of
notification had to be manually sent to the user to inform them of the
need to use a new username when logging in.