Make project upload path non-optional again
When creating a new project, the project upload path is meant to be
optional (it can be left blank). There was a bug where a trailing slash
was added to a path that wasn't blank. Then a check was performed to see
if the path was blank, and if not, the path was checked for validity.
Obviously the path would never be blank, because it'd always contain a
trailing directory separator.