Show default upload path when creating project
Implementation of bug #8250

When creating a project, the user has an option to specify a separate
upload folder on a per-project basis. Previously this field was left
blank by default. This patch shows the default value to the user - as
long as they're an administrator. This helps make it a little easier to
see if a new path needs to be specified that is different from the
default. For security reasons the absolute upload path is only shown to
administrators.