Don't update target_version if no permission
It isn't possible to set a blank target_version on a bug object if the
user hasn't got permission to do so (roadmap_update_threshold).
Therefore, when reporting bugs, the target_version on the new bug object
should not be set to a blank value as an error will be thrown in
__set complaining that the user hasn't got permission to change
target_version.