Option to show scheduled version release dates
This is a new feature which allows the scheduled release dates for
versions to be shown next to product version, target version and fixed
in version fields. Disabled by default. Enable this feature by setting
$g_show_scheduled_release_dates = ON in your configuration. Showing
release dates can be quite useful for quickly seeing how far away a
target version is due, etc.

Also renamed instances of $t_show_verison to $t_show_product_version to
reduce confusion in the codebase as to whether it is referring to the
Mantis version ($g_show_version) or the product versions
($g_show_product_version).

Resolves bug #10274.