Add option to hide projection field
Fixes bug #10618.

We've had options $g_enable_eta and $g_enable_product_build for some
time now. This patch adds a similar option to hide display of the
projection field. The aim of this patch was to match the implementation
of the options for the eta and product build fields as closely as
possible.

Ideally in the future, a better solution needs to be created for
disabling or adjusting the settings of these non-essential fields.
Possibly moving to custom fields would be a better idea. I consider this
patch an interim fix.