Anonymous accounts: fixes and doc updates
Updated documentation of configuration options relating to anonymous
access to Mantis. Fixes #10626.

Added new user_is_anonymous function to the user API. The purpose of
this function is to check if the current user is the anonymous user
account.

Changes user_is_protected function in the user API to always return true
when the user is the anonymous user account. The anonymous user should
never be allowed to update the preferences and settings on the anonymous
account! Fixes #10265.

Fixed a major security bug in account_prefs_reset.php that allowed
anyone to reset the account preferences of another person. Fixes #10264.