Deprecate use of custom_constant_inc.php
This is a follow-up to 1ae2c4fb08701e5e4f0a47bc6078edd53797bef2 where
custom_constant_inc.php was renamed to custom_constants_inc.php for
naming consistency.

The 1.2 release cycle gives users the opportunity to migrate to the new
name of this file in preparation for the release of version 1.3.