CSRF protection not needed for bug_report form
bug_report.php has a redirect form that sends a copy of details from the
last bug reported to a new bug report form. This makes it easy to create
a bunch of similar issues in a row. CSRF is not required here because it
doesn't result in changes being made to Mantis.