Remove offtopic uses of handle_bug_threshold
handle_bug_threshold was being used beyond it's scope, in an
inconsistent way, to prevent users updating certain bug fields.

This change exposes the priority field to the advanced bug report page.
Anyone that can report a bug can update this field. In fact, users could
always update the priority field by generating their own request to
bug_report.php. The only thing that has changed is that the UI has been
updated to match what happens behind the scenes.

If we want to limit who can read/write to the priority field (and other
fields) we really need to create some new thresholds that are checked in
bug_api.php within __set, and elsewhere throughout the code as
appropriate.