Fix #10714: Fix session notice with verify.php
This problem is a result of changes for issue #10187, where the session
is no longer destroyed in order to play nice with other PHP
applications.  However, the problem itself only manifests itself when a
user that is already logged into mantis follows the verification link.

Rather than forcefully restarting the session, the verification page now
logs out the existing user and then does a header redirect to itself to
allow the normal session init behavior to kick in.