Issue #10696: more customisation of version showing
In some installations, the product version field may not be used.
However, the target version and fixed in version fields may still be
used in such cases.

An example is where users report bugs or issues where they don't know
what version the bug occurred in. Versions may actually be defined to be
the week number of the year. The target version may still be used in
these cases to let users know what week a resolution can be expected
by.