Fix #3917: Ability to delete other users from bug monitor list
A new configuration option $g_monitor_delete_others_bug_threshold was
added. This option defines the access level required to delete other
users from the list of users monitoring a bug.

If the current user is in an access level above or equal to
$g_monitor_delete_others_bug_threshold they will see a [Delete] link
next to each username in the bug monitor list for each bug. Clicking on
this link will stop a user from monitoring a bug.