Fix #10746: view issue range incorrect on first view
When an anonymous user or someone else without a saved filter state
views the bug list for the first time, current_user_get_bug_filter()
will fail. This led to the bug view range not working (as it didn't know
how many bugs are being shown per page).

In such cases where a per-user filter is not yet known, we should use
the default filter. Note that this check may be better placed in
current_user_get_bug_filter() if it is deemed safe to do so (it changes
the output of the function).