Fix #10687: Force use of account_delete when deleting own account
The case of deleting ones own account is quite different to deleting the
account of another user. Therefore if an administrator wants to delete
their own account, account_delete.php should be used instead. It
correctly handles logging out and redirection of the administrator who
has just deleted their own account.

This fix will force account_delete.php to be used in a way that is
transparent to an administrator who is deleting their account.

For the purpose of this commit message, "administrator" is any user who
has an access level equal to or beyond manage_user_threshold.