Fix #10815: hide some internal fields from columns_api
It is possible through account_manage_columns_page.php (and elsewhere)
to view columns such as '_stats', 'profile_id', 'sticky' and 'loading'
which are either just internal bits of data (not intended to be seen by
the user) or in the case of 'sticky', are for internal display purposes
only.

These unnecessary columns should not be exposed through columns_api.