Issue #10859: Add hyperlinks for bugnote IDs
To make it easier to reference bugnotes externally (in emails, IRC,
etc), bugnote IDs on the bug view page should be hyperlinked to match
the existing anchor for each bugnote.

Note that bug_view_inc.php is included by other scripts (such as when
one attempts to update a bug). For this reason, view.php is always used
when creating the bugnote hyperlink, as it doesn't make sense to send an
external user to a bug update page, just so they can view a bugnote.

Signed-off-by: David Hicks <hickseydr@optusnet.com.au>