Issue #6047: Use finfo PHP functions to determine file types
The finfo class is now standard in PHP 5.3.0 and can be used to
determine the MIME type of attachments. It is best to perform this check
upon each download so that updates in PHP and the libmagic library
behind the scene will apply to all attachments - not just ones created
after the update.