db_field_exists fails when no column names are returned by AdoDB
This commit fixes up the error handling within db_field_exists so that
the function returns false if AdoDB cannot return any column names from
a table in the database. Previously an error was raised where a
comparison using in_array() has an invalid type for the second
parameter. This is because the MetaColumnNames function either returns
an array, or false (boolean).