Fix #10914: Make db_get_table behave like plugin_table
db_get_table is currently called with a parameter in the format
"mantis_{tablename}_table". This is unnecessary. The function should
behave more like plugin_table where the name of a table such as "bug" or
"email" is passed in, and the optional user-customisable prefix and
suffix are prepended and appended.

This change removes some processing overheads and makes the function a
little bit easier to use, as it is consistent with the usage for
plugin_table.

Note that this breaks the API and therefore plugins will most likely
have to change the parameters they're sending to db_get_table.