Issue #10967: Check for existence of ini_set function before use
On some servers, the administrator may have disabled the use of the
ini_set function for security reasons (using the disabled_functions PHP
configuration option).

We should check to make sure ini_set is available before we try to call
it.

Also use set_magic_quotes_runtime() instead of doing an ini_set on the
"magic_quotes_runtime" option. This function is deprecated as of PHP
5.3.0 but at least we can do a function_exists() check on it.