Complete move of graphing abilities to MantisGraph
The move of the graphing abilities from the core into a new MantisGraph
plugin wasn't fully completed. This large patch completes the move by
ripping out remaining parts of graphing support from the core, and
moving them into the plugin.

Some of the things that have changed as a result of this commit:

1. Deprecated the graph_font configuration option in the core and added
it back to the plugin configuration. This time there is a nicer UI for
selecting the font from within the plugin configuration screen.

2. Deprecated the graph_colors option from the core and disabled this
functionality in the plugin. This will obviously require a lot more work
(outside the scope of what I can do at the moment) so that there is a
better way for the user to select colours via the UI configuration page.
Additionally, this option shouldn't just be restrictied to JpGraph.
Graph colours need to be reimplemented properly at a later stage.

3. Deprecated the default_graph_type option from the core and disabled
this functionality in the plugin. This was a very cryptic non-documented
option that set the default plot type on bug_graph_page. If this is to
be reimplemented, it needs to be done so properly so that it isn't
cryptic and unusable. There is no real loss of functionality with the
removal of this option.

4. Reordered the layout of configuration options in the plugin
configuration page so similar options are grouped together. It also
looks much nicer now thanks to some changes to strings, and the
placement of comments on a second line in small text.

5. Moved all strings used solely by MantisGraph into the plugin. The
original strings have been removed from the core as they aren't used
there.

6. Removed references to the graph related options from the
Configuration section of the Administrator Guide.

7. Moved class.period.php from the core to the MantisGraph plugin as it
currently isn't used elsewhere within Mantis. Really this class should
be part of the core of Mantis (under
core/classes/MantisPeriod.class.php) but it'll need some more work
before it can be used in other parts of Mantis. For instance, we don't
use this class for date filters. More importantly, PHP has a built-in
class called DatePeriod (and DateInterval) that appear to be better
choices for future work with date periods.

8. Added a UI field in the plugin configuration page that is only
accessible to users with access level >= admin_site_threshold to set
what once used to be $g_jpgraph_path.