Fix #10748: Fix due_date column, add new overdue column
This patch implements support for showing the due_date field as a
column.

Additionally, the overdue icon has been moved from within the edit
column to it's own new "overdue" column.

The functionality of the due_date and overdue columns do overlap, so it
only makes sense to use one at a time.