Allow specification of magic database file for fileinfo
This patch adds a new configuration option 'fileinfo_magic_db_file'
which allows an administrator to specify the filename of the magic
database file used by the PHP fileinfo extension (now part of 5.3.0 by
default).

Some installation documentation has also been included to let
administrators know that it is a good idea to make fileinfo available
for Mantis to use.