Issue #10998: Add missing EVENT_BUGNOTE_ADD_FORM event to extra pages
There is an ability to add a bugnote on both the change status and bug
update pages. However, EVENT_BUGNOTE_ADD_FORM was not being raised as an
event in these instances. This meant that plugins could only add extra
bugnote fields on the form printed to the view page. Plugins should be
able to print extra fields to all bugnote forms.