Fix #10703: Don't show attachment_count column in CSV output
It is a little complicated to show the attachment_count column in the
CSV output at the moment. Therefore we disable this column from being
usable until such time as someone wants to spend time fixing it up
properly. By disabling the column, we prevent users from enabling what
is currently a broken/non-implemented column.