Issue #7917: "Update Information" button at top of update form
This patch from stappel extends the use of the existing
$g_action_button_position configuration option to apply to the "Update
Information" button on the bug update form. This can make it a little
bit quicker/easier to update a bug without scrolling all the way down
the page to reach the bottom of the form. Again, this is optional and
disabled by default.