Fixes #10592: mc_enum functions doesn't take customization into account
The mc_enum_xxx functions have been updated to return the customised
keys, while the names remain the translated ones.

mc_enum_get has not been updated, as it returns a raw enum string, not
an ObjectRefArray.

Unfortunately no new tests can be added since we don't control the
customisation settings on the mantis installation on a per-test
basis.

- split functionality from mc_issue_get
- explode_bla_bla now takes a enumeration name
- localized labels are properly used