Fix #10754: Allow user to enter redirect delays above 9 seconds
Users were prevented from entering redirect delays larger than 9 because
the field which accepted the redirect delay was limited to a single
character of input.

I've changed the allowable number of characters in the input field to 3
to allow for redirect delays above 2 minutes. Users shouldn't have a
need to set the redirect delay larger than 40 minutes (the maximum this
field can take with just 3 characters allowed in the field). However
it's quite reasonable for a user to expect a redirect delay of 2 or 5
minutes.