Allow SOAP tests to delete issues on tearDown
Until now deletion of issues created in tests was done manually, which
had the disadvantage of not being executed unless the test did not throw
an exception. Then the test fails, the issue is not deleted.

This commit introduces a deleteAfterRun method in the SoapBase class
which allows tests to register a issue id to be deleted. The issues
are then deleted on tearDown, which runs even if an assertion has
failed when the test is executed.