Use string os_version for platform version
When dealing with platform data, language string 'os_version' should
be preferred over the normal 'version' for clarity of meaning, and to
allow sites/plugins to override the meaning of one without affecting the
other context. In this case, overriding 'version' for project contexts
would affect the platform contexts when users were editing their set of
profiles.