Return a proper soap_fault if attachment retrieval fails
Attempts to get an inexisting attachment had inconsistent results,
depending on the attachment_type, either returning empty for
project attachents or throwing an internal error ( no bug with
id 0 ).

This commit changes the API to return a soap_fault is the
attachment can not be retrieved and verifies it with two
new tests.