Use floating spans to better position page links
When adding custom columns, the old method of using colspan to place
page links on the right side breaks down, often resulting in page links
broken over multiple lines, with plenty of whitespace to the left.

Rather than trying to add an intelligint column counting algorithm, the
colspan was replaced with a span floated to the right side.  We'll leave
it up to the browser to handle overflow.