Implemented method for columns to pre-cache data
When implementing a custom column, there needs to be a method for the
column to pre-fetch and cache data related to the set of viewed issues,
rather than requiring individual queries to be made for each issue in
the dataset.

A basic implementation would be to override the ->cache() method to
generate a single query to load and store that data in a static property
that can be accessed by the ->display() method.