Implement optional sorting for custom columns
In order to allow custom columns to be useful and consistent with native
columns, the user must be able to sort issues based on the custom
columns.  An optional sorting implementation was added based on the
method for adding custom filters to the filter queries.  A flag was
added to the MantisColumn archetype to allow the column author to
specify if the column is sortable; if not, the column header will not be
clickable, preventing user confusion.